LayoutTests/fast/shadow-dom/fullscreen-in-shadow-full-screen-ancestor-expected.txt

Test that webkitFullscreenElement retargets the fullscreen element inside a shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector(":-webkit-full-screen-ancestor") is document.documentElement
PASS successfullyParsed is true

TEST COMPLETE
Enter fullscreen
